irio/requirements.txt
pandas==2.0.1
python-Levenshtein==0.21.0
openai==0.27.7